requirements.txt
streamlit==0.74.1
pyodbc==4.0.30
Pillow==8.1.0